При вызове fetchPhotos вынес cb в отдельные функции
